+++ src/structs.h
@@ -6,14 +6,15 @@
 #define WAIT_TIME 15
 
 typedef struct fcli1{ // <from cliente>
-    long long size;
+    // use unsigned short
+    unsigned short size;
     char nome[50];
     pid_t pid_cliente; // maneira de o identificar
     char sintomas[TAM_MAX_MSG]; //sintomas do cliente
 }sint_fcli, *psint_fcli;
 
 typedef struct fblc1{ // <from balcao>
-    long long size;
+    unsigned short size;
     char esp[20]; // <especialidade>
     int prio; // <prioridade>
     int num_peopleAhead; // num de especialistas à frente nessa especialidade
@@ -21,7 +22,7 @@ typedef struct fblc1{ // <from balcao>
 }info_fblc, *pinfo_fblc;
 
 typedef struct fmed1{ // <from medico>
-    long long size;
+    unsigned short size;
     char nome[50];
     pid_t pid_medico; // maneira de o identificar
     char esp[20];
@@ -29,12 +30,12 @@ typedef struct fmed1{ // <from medico>
 }esp_fmed, *pesp_fmed;
 
 typedef struct fblc2{
-    long long size;
+    unsigned short size;
     char nome[50];
 }connectMed_fblc, *pconnectMed_fblc;
 
 typedef struct fblc3{
-    long long size;
+    unsigned short size;
     char nome[50];
     pid_t pid_cliente;
 }connectCli_fblc, *pconnectCli_fblc;
